split circle gizmos out into their own file
