+++ crates/bevy_gizmos/src/gizmos.rs
@@ -2,6 +2,7 @@
 
 use std::{f32::consts::TAU, iter};
 
+use crate::circles::DEFAULT_CIRCLE_SEGMENTS;
 use bevy_ecs::{
     system::{Deferred, Resource, SystemBuffer, SystemMeta, SystemParam},
     world::World,
@@ -12,8 +13,6 @@ use bevy_transform::TransformPoint;
 
 type PositionItem = [f32; 3];
 type ColorItem = [f32; 4];
-
-const DEFAULT_CIRCLE_SEGMENTS: usize = 32;
 
 #[derive(Resource, Default)]
 pub(crate) struct GizmoStorage {
@@ -199,44 +198,6 @@ impl<'s> Gizmos<'s> {
 
         strip_positions.push([f32::NAN; 3]);
         strip_colors.push([f32::NAN; 4]);
-    }
-
-    /// Draw a circle in 3D at `position` with the flat side facing `normal`.
-    ///
-    /// This should be called for each frame the circle needs to be rendered.
-    ///
-    /// # Example
-    /// ```
-    /// # use bevy_gizmos::prelude::*;
-    /// # use bevy_render::prelude::*;
-    /// # use bevy_math::prelude::*;
-    /// fn system(mut gizmos: Gizmos) {
-    ///     gizmos.circle(Vec3::ZERO, Vec3::Z, 1., Color::GREEN);
-    ///
-    ///     // Circles have 32 line-segments by default.
-    ///     // You may want to increase this for larger circles.
-    ///     gizmos
-    ///         .circle(Vec3::ZERO, Vec3::Z, 5., Color::RED)
-    ///         .segments(64);
-    /// }
-    /// # bevy_ecs::system::assert_is_system(system);
-    /// ```
-    #[inline]
-    pub fn circle(
-        &mut self,
-        position: Vec3,
-        normal: Vec3,
-        radius: f32,
-        color: Color,
-    ) -> CircleBuilder<'_, 's> {
-        CircleBuilder {
-            gizmos: self,
-            position,
-            normal,
-            radius,
-            color,
-            segments: DEFAULT_CIRCLE_SEGMENTS,
-        }
     }
 
     /// Draw a wireframe sphere in 3D made out of 3 circles around the axes.
@@ -466,42 +427,6 @@ impl<'s> Gizmos<'s> {
         self.line_gradient_2d(start, start + vector, start_color, end_color);
     }
 
-    /// Draw a circle in 2D.
-    ///
-    /// This should be called for each frame the circle needs to be rendered.
-    ///
-    /// # Example
-    /// ```
-    /// # use bevy_gizmos::prelude::*;
-    /// # use bevy_render::prelude::*;
-    /// # use bevy_math::prelude::*;
-    /// fn system(mut gizmos: Gizmos) {
-    ///     gizmos.circle_2d(Vec2::ZERO, 1., Color::GREEN);
-    ///
-    ///     // Circles have 32 line-segments by default.
-    ///     // You may want to increase this for larger circles.
-    ///     gizmos
-    ///         .circle_2d(Vec2::ZERO, 5., Color::RED)
-    ///         .segments(64);
-    /// }
-    /// # bevy_ecs::system::assert_is_system(system);
-    /// ```
-    #[inline]
-    pub fn circle_2d(
-        &mut self,
-        position: Vec2,
-        radius: f32,
-        color: Color,
-    ) -> Circle2dBuilder<'_, 's> {
-        Circle2dBuilder {
-            gizmos: self,
-            position,
-            radius,
-            color,
-            segments: DEFAULT_CIRCLE_SEGMENTS,
-        }
-    }
-
     /// Draw an arc, which is a part of the circumference of a circle, in 2D.
     ///
     /// This should be called for each frame the arc needs to be rendered.
@@ -603,33 +528,6 @@ impl<'s> Gizmos<'s> {
     }
 }
 
-/// A builder returned by [`Gizmos::circle`].
-pub struct CircleBuilder<'a, 's> {
-    gizmos: &'a mut Gizmos<'s>,
-    position: Vec3,
-    normal: Vec3,
-    radius: f32,
-    color: Color,
-    segments: usize,
-}
-
-impl CircleBuilder<'_, '_> {
-    /// Set the number of line-segments for this circle.
-    pub fn segments(mut self, segments: usize) -> Self {
-        self.segments = segments;
-        self
-    }
-}
-
-impl Drop for CircleBuilder<'_, '_> {
-    fn drop(&mut self) {
-        let rotation = Quat::from_rotation_arc(Vec3::Z, self.normal);
-        let positions = circle_inner(self.radius, self.segments)
-            .map(|vec2| (self.position + rotation * vec2.extend(0.)));
-        self.gizmos.linestrip(positions, self.color);
-    }
-}
-
 /// A builder returned by [`Gizmos::sphere`].
 pub struct SphereBuilder<'a, 's> {
     gizmos: &'a mut Gizmos<'s>,
@@ -655,30 +553,6 @@ impl Drop for SphereBuilder<'_, '_> {
                 .circle(self.position, self.rotation * axis, self.radius, self.color)
                 .segments(self.circle_segments);
         }
-    }
-}
-
-/// A builder returned by [`Gizmos::circle_2d`].
-pub struct Circle2dBuilder<'a, 's> {
-    gizmos: &'a mut Gizmos<'s>,
-    position: Vec2,
-    radius: f32,
-    color: Color,
-    segments: usize,
-}
-
-impl Circle2dBuilder<'_, '_> {
-    /// Set the number of line-segments for this circle.
-    pub fn segments(mut self, segments: usize) -> Self {
-        self.segments = segments;
-        self
-    }
-}
-
-impl Drop for Circle2dBuilder<'_, '_> {
-    fn drop(&mut self) {
-        let positions = circle_inner(self.radius, self.segments).map(|vec2| (vec2 + self.position));
-        self.gizmos.linestrip_2d(positions, self.color);
     }
 }
 
@@ -726,13 +600,6 @@ fn arc_inner(
         let start = direction_angle - arc_angle / 2.;
 
         let angle = start + (i as f32 * (arc_angle / segments as f32));
-        Vec2::from(angle.sin_cos()) * radius
-    })
-}
-
-fn circle_inner(radius: f32, segments: usize) -> impl Iterator<Item = Vec2> {
-    (0..segments + 1).map(move |i| {
-        let angle = i as f32 * TAU / segments as f32;
         Vec2::from(angle.sin_cos()) * radius
     })
 }
