+++ crates/bevy_gizmos/src/lib.rs
@@ -17,6 +17,7 @@
 //! See the documentation on [`Gizmos`] for more examples.
 
 mod arrows;
+mod circles;
 pub mod gizmos;
 
 #[cfg(feature = "bevy_sprite")]
